Changes done if user changed default avatar

+++ app/Repositories/UserRepository.php
@@ -51,7 +51,10 @@ class UserRepository {
         }
         
         if($request->hasFile('avatar')){
-            unlink($user->avatar);
+            if(!$user->avatar === 'images/avatar/default.jpg')
+            {
+                unlink($user->avatar);
+            }
             $avatar = $request->file('avatar');
             $filename = time() .'.'. $avatar->getClientOriginalExtension();
             $photo = Image::make($avatar)->resize(300, 300)->save(public_path('images/avatar/'.$filename));
